# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DreamCar Reserva de vehiculos",
        "description": "Esta API nos permitira ver y manipular informacion a las reservas de los vehiculos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vehiculos/show/{id}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Obtener un vehiculo por id",
                "description": "Renorna los detalles de un vehiculo",
                "operationId": "719d045fc13b96101c054f7526251e4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehiculo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del vehiculo con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "id_marca": {
                                            "type": "integer",
                                            "example": "99"
                                        },
                                        "modelo": {
                                            "type": "integer",
                                            "example": "2010"
                                        },
                                        "precio": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "disponible": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehiculos/vehiculo/{id_marca}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Buscar vehiculo por id de una marca",
                "description": "Retorna un vehiculo",
                "operationId": "2105ec78ee054e6a77ca26c7360c597b",
                "parameters": [
                    {
                        "name": "idMarca",
                        "in": "path",
                        "description": "id de la marca relacionada al vehiculo a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "vehiculo": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "id de una Marca",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la marca del vehiculo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehiculos/vehiculo/{modelo}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Buscar vehiculo por modelo",
                "description": "Retorna un vehiculo",
                "operationId": "e3569ca0974fe4319e08e65f140df06d",
                "parameters": [
                    {
                        "name": "modelo",
                        "in": "path",
                        "description": "Modelo del vehiculo a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modelo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "modelo": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Modelo de un vehiculo",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro el modelo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}